\name{psr}
\alias{psr}
\title{
psr
}
\description{
Constructs a penalized spline regression of the historical time series data to be used in flow field forcasting.
}
\usage{psr(t, y)}

\arguments{
  \item{t}{
Time series observation times.
}
  \item{y}{
Time series response values.
}
}
\value{
Returns the penalized spline regression data skeleton.
}
\references{
D. Ruppert, M. P. Wand and R. J. Carroll, Semiparametric Regression. 
New York, NY: Cambidge University Press, 2003.
}
\author{
Kyle A. Caudle
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
